Fix 128-bits type definition

+++ fixedpoint_config.h
@@ -3,8 +3,8 @@
 
 #ifdef __x86_64__
     #define FRACT_HAS_128BITS
-    typedef unsigned long long uint128_t;
-    typedef signed long long int128_t;
+    typedef __uint128_t uint128_t;
+    typedef __int128_t long int128_t;
 #endif
 
 // Avoid using any division
